+++ yarn-project/foundation/src/serialize/serialize.test.ts
@@ -1,5 +1,5 @@
 import { randomBytes } from '../crypto/index.js';
-import { Fr } from '../index.js';
+import { Fr } from '../fields/index.js';
 import {
   serializeBufferToVector,
   deserializeBufferFromVector,
